+++ app/Http/Controllers/InstitutoController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\InstitutoStoreRequest;
+use App\Http\Requests\UpdateInstitutoRequest;
 use App\Models\Carrera;
 use App\Models\Instituto;
 use Exception;
@@ -41,18 +43,20 @@ class InstitutoController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(InstitutoStoreRequest $request)
     {
+        try {
+            $this->authorize('crearInstituto', App\Models\Instituto::class);
+            $instituto = new Instituto();
+            $instituto->nombre_instituto = $request->nombre;
+            $instituto->save();
 
-        $instituto = new Instituto();
-        $instituto->nombre_instituto = $request->nombre;
-        $instituto->save();
+            $instituto->carreras()->attach($request->input('ids'));
 
-        foreach ($request->input('ids') as $idCarrera) {
-            $instituto->carreras()->attach($idCarrera);
+            return redirect()->route('institutos.index');
+        } catch (Exception) {
+            return redirect('@me');
         }
-
-        return redirect()->route('institutos.index');
     }
 
 
@@ -69,28 +73,38 @@ class InstitutoController extends Controller
      */
     public function edit(Instituto $instituto)
     {
+        try {
+            $this->authorize('editarInstituto', App\Models\Instituto::class);
 
-        $carrerasI = $instituto->carreras;
-        $carreras = Carrera::whereDoesntHave('institutos', function ($query) use ($instituto) {
-            $query->where('instituto_id', $instituto->id);
-        })->get();
+            $carrerasI = $instituto->carreras;
+            $carreras = Carrera::whereDoesntHave('institutos', function ($query) use ($instituto) {
+                $query->where('instituto_id', $instituto->id);
+            })->get();
 
-        return view('admin.institutos.editar', compact('carreras','carrerasI', 'instituto'));
+            return view('admin.institutos.editar', compact('carreras', 'carrerasI', 'instituto'));
+        } catch (Exception) {
+            return redirect('@me');
+        }
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Instituto $instituto)
+    public function update(UpdateInstitutoRequest $request, Instituto $instituto)
     {
-        $instituto->nombre_instituto = $request->nombre;
-        $instituto->save();
+        try {
+            $this->authorize('editarInstituto', App\Models\Instituto::class);
+            $instituto->nombre_instituto = $request->nombre;
+            $instituto->save();
 
-   
-        $instituto->carreras()->sync($request->input('ids'));
-  
 
-        return redirect()->route('institutos.index');
+            $instituto->carreras()->sync($request->input('ids'));
+
+
+            return redirect()->route('institutos.index');
+        } catch (Exception) {
+            return redirect('@me');
+        }
     }
 
     /**
@@ -98,7 +112,13 @@ class InstitutoController extends Controller
      */
     public function destroy(Instituto $instituto)
     {
-        $instituto->delete();
-        return redirect()->route('institutos.index');
+        try {
+            $this->authorize('eliminarInstituto', App\Models\Instituto::class);
+            $instituto->delete();
+            
+            return redirect()->route('institutos.index');
+        } catch (Exception) {
+            return redirect('@me');
+        }
     }
 }
